[Post Message] Block Same-Site Cross-Partition postMessage
This hasn't launched so it's defaulting off, but we need to get the
new test and expectation live because it includes a fix to data
gathering in DocumentLoader. With this fix, we can properly figure
out if it's safe to block same-origin different-partition post messages.

Bug: 1159586
Change-Id: Ia97aa174fb325d0ce367a945237459cbbb88ec28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3891594
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055221}