Prevent clearing InkDrop state when context menu is shown for BookmarkButton
Setting this flag to false prevents the Button class from clearing the InkDrop state when a context menu is shown.

Note that this does not completely fix the bug, but instead uncovers an more confounding issue. That is, after (and only after) doing some specific actions in the BookmarkBar, the BookmarkButton InkDrop state will work as expected. I'll add these reproductions and other details to the linked bug.


Bug: 1350143
Change-Id: I2a08d0debd514585e5fe3b1387867ff1b3e7f3f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3929482
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Mickey Burks <mickeyburks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055259}