Plumb parent's BaseURL value to child renderer for srcdoc iframe.
This CL modifies the plumbing that sends 'srcdoc' attribute values
from the parent frame's renderer process to the child frame's
renderer process so that it also sends the parent's
Document::BaseURL() and makes it available in the child document's
FallbackURL(). The parent frames's RenderFrameHost maintains an
accurate view of each frame's base URL at all times, and this is
sent to a child frame if it navigates to srcdoc.

This is necessary so that srcdoc iframes that are sandboxed, which
put the child in a different renderer process, still have access to
the same BaseURL() as if it was in the same renderer process.

These changes presently only take effect when the
IsolateSandboxedIframes is enabled, in which case the base URL
behavior changes to proper snapshotting per Issue 1356658 (whether
it's out of process or not).

Bug: 1339824, 1343065, 1356658
Change-Id: I0d1d9ef05121bd603a3f60bcdc7181f65a37f77b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3792049
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055296}