Add TraceEvent calls to WebContentsAccessibilityImpl
This is a small CL which adds some calls to TraceEvent begin/end to
the Java-side Android accessibility code.

This will make future tracing and/or debugging easier.
Using these traces we can ensure that future performance and
accessibility refactoring work is not resulting in unnecessary
construction/destruction of the WebContentsAccessibilityImpl
objects.

AX-Relnotes: N/A
Bug: N/A
Change-Id: I3940dba24c856354fa4dd7efa5a140d2a7265891
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3931065
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1055312}