[Android Bulk Tab Edit] Fix group bugs
Fix UI update issues with the TabSelectionEditorGroupAction when
performed on existing groups or undoing aggregation of multiple groups.

The root cause is that the sequence of didMoveTabOutOfGroup and
didMergeTabToGroup calls to the observer in TabListMediator were not
valid. This resulted in the data model in TabGroupModelFilter being
correct, but the TabListMediator displayed incorrect information
until forcibly reset.

Description of the fix:

1) In undo every tab signals didMoveTabOutOfGroup. This CL updates
   TabListMediator to not create a new tab card in the grid/list if the
   tab has related tabs in the TabGroupModelFilter (i.e. it is part of a
   group). With this change, only the first tab for a restored group
   will create a new tab card. This is called for every tab to correctly
   remove each tab from the multi-thumbnail of the previous tab card.

2) For the first restored tab didMoveTabOutOfGroup will create a card &
   nothing else needs to be done. The remaining tabs need to be moved
   to the card created for the first tab. This is achieved by signalling
   didMergeTabToGroup for every tab other than the first. In
   TabListMediator's implementation the desIndex will be valid, but
   srcIndex will be -1 as from the perspective of TabListMediator this
   operation moves a tab between groups so there is no card to merge to
   the destination. didMergeTabToGroup now special cases when this
   happens to just update the tab at desIndex with the correct selection
   state and update the multi-thumbnail.

The last selected tab in the Restored Group will be lost information.
We can fix this in the TabListMediator UI by marking the tab as selected
if it is the selected tab in the TabModel. However, for groups not
containing the selected tab the last selected tab will default to the
first tab that was restored to the group.

Bug: 1343374
Change-Id: I30430acd8183499fc47f6750a54e561e7a4d7f19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3864078
Reviewed-by: Mei Liang <meiliang@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055338}