Reorder some code in RealtimeAnalyser
While working on a different bug, I noticed that the order of some
code in RealtimeAnalyser isn't compliant with the coding specification.

This CL only changes the order of code blocks and should have no
functional change.  It is to fix some minor ordering issues, and
to prepare for later refactoring.

Change-Id: I57b03a86e2f733ccd32312e9a643907e9c352214
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3933272
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Michael Wilson <mjwilson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055396}