Skip Creating Template URL Service Sometimes
It doesn't need to be created for system profile or for most of the
unusual ChromeOS profiles.  None of those can be used to do searches.
The only exception is the "signin profile" for ChromeOS, which does need
a template URL service, which is used to create and display the URL
bar.

In the process,
* rename the histogram Search.DefaultSearchProviderType.
This is the only histogram as far as I can tell that's emitted when
a template URL is constructed.  Replace it with a version 2 and mark
the original as obsolete.
* change the conditions under which NavEntryCommitted.SRP is emitted.  (Chrome shouldn't try to emit it in profiles where the user isn't allowed to do searches on their own.)

Bug: 1284659
Change-Id: Ifd0a130de423da4afa7fac5984db8ac019257280
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3913897
Commit-Queue: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055430}