[Read Anything] Redefine lang attribute setting
Currently, for every Read Anything node, we call ax_node's GetLanguage,
which walks up the tree to find the language attribute.

In this CL, instead of walking up the tree for every node, we only call
GetLanguage on the root nodes in Read Anything. For all other nodes, we
get the lang attribute from the corresponding node in the ax tree. This
way we are not walking up the tree more than necessary and only certain
parent nodes have the language set.

Bug: 1266555
Change-Id: Ia2bbbd462f0faf3ab1cc1f42aa106c07eff43116
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3932084
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Commit-Queue: Jocelyn Tran <jocelyntran@google.com>
Cr-Commit-Position: refs/heads/main@{#1055433}