SET: Use frame size for root image intrinsic sizes
This CL changes the SET container pseudo to be the size of the viewport
"as-if" all viewport-insetting widgets (URL bar, keyboard, etc.) were
hidden. The container's style is set so that, if the viewport is insetted, the container is shifted up and to the left, underneath the
UI widets, to appear where it would if all UI widgets were hidden.

This ensures the size and screen-relative position of the container
is always stable across a transition, regardless of the state of UI
controls before and after the transition begins. E.g. consider a
navigation; the snapshot is taken with the URL bar hidden but the
navigation forces the URL bar to be shown. This change ensures the
container can appear fixed across these two states.

To account for this change, Blink painting code expands the ViewRect
of the LayoutView so that the produced render pass matches the size
of the container. Additionally, painting of all content descending
from the LayoutView is offset by the viewport-insets so that it paints
at the visual (0, 0) origin of the actual viewport (which is not at
origin of the container).

Bug: 1348555
Change-Id: Ib7534f15d368da2a4cd3147ce59a6f93167ae620
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3863091
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Auto-Submit: David Bokan <bokan@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055442}