[Fast Pair] Prevent notifications for devices lost in Mediator
When the Mediator receives a device lost event, trigger the timeout
through the UiBroker to the FastPairPresenter to prevent notifications
from appearing for the device for the timeout. This change also adds
a flag for triggering the timeout.

TEST=
Verified on DUT that flag enabled : devices lost don't reappear for 30 seconds, or until Bluetooth is toggled, user signs out, or Fast Pair is toggled, and flag disabled: no timeout for devices lost.

Bug: b/244472452
Change-Id: I5493703ef0f11a67bc09a20c9bd137d9a6750784
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3923743
Commit-Queue: Juliet Lévesque <julietlevesque@google.com>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055513}