Remove FPS-enabled plumbing in the network stack.
Since we already have the FPS-enabled bit in
FirstPartySetsAccessDelegate, and those instances are already the way
that NetworkContexts access FPS info, we don't need to duplicate this
bool elsewhere in the network service. Furthermore, since this bool
is going to become mutable, updating it in only one place is much
simpler than updating all of the copies of it (and hoping that there
aren't subtle inconsistencies as a result).

Bug: 1286070, 1366918
Change-Id: I7e4700a2619951e1b9950e1a9811c9cbbe4f74f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3934021
Auto-Submit: Chris Fredrickson <cfredric@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055556}