[Chrome/Android] Refactor to current drag/Drop feature
In order to add the drag/drop feature for Webview we will need to make a couple of things:

We extract the core implementation from DropDataContentProvider to a new class DropDataProviderImpl, so it can be shared with WebView;
Also, we changed the way we access DropDataContentProvider into replying on Context#getContentResolver. This is needed for WebView, as the WebView is created via a different class loader and have no direct access to DropDataContentProvider class.


Bug: 1371557
Change-Id: I971c3435171adbc93c25ab333dcfe01bcfda4f6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3909341
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055584}