Update Shim_TestRulesRegistryIDAreRemovedAfterWebViewIsGone test
This CL updates Shim_TestRulesRegistryIDAreRemovedAfterWebViewIsGone
test from WebViewChannelTest based on the current code and enables
it.

The test was disabled on [1]. When the test worked,
GetRulesRegistry() returns null if it doesn't find a specified
RulesRegistry.
On the current code base, in that case, it returns a new
RulesRegistry. So, it's not proper to use to check if
the RulesRegistry exists.
So, this CL introduces a new helper, HasRulesRegistryForTesting(),
for that purpose and updates the test to use it.

This CL also adds content::ScopedAllowRendererCrashes to
Shim_TestRulesRegistryIDAreRemovedAfterWebViewIsGone test since the
test kills a renderer intentionally.

[1] https://codereview.chromium.org/768743004

Bug: 438327
Change-Id: I1acfbe77cdc3a3c9fb9679150d96e726a0636bd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3925640
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1055647}