[fyfre]Split sign in page files of Lacros and Dice
The FRE for Lacros and Dice won't be the same so
we decided to split the HTML and Typescript files
before moving on. We needed to do this because in future
changes, when adding the strings to be translated in the .grd files,
we need to guard them for Lacros.

Bug: 1347507
Change-Id: Ifaff861fb504a537ef38acebaa782069df9a824c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3930918
Reviewed-by: Monica Basta <msalama@chromium.org>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: Jack Yammine <jyammine@google.com>
Cr-Commit-Position: refs/heads/main@{#1055686}