Deduplicate file events (timed-entries version)
The procedures where we perform file reporting are called several times
for a single user-initiated action. The number of calls to these
procedures might be unpredictable (e.g., third-party app trying without
success to open a confidential file for a certain number of attempts).
This CL implements a simple deduplication strategy based on reporting
only if enough time has passed from a previous reporting of the same
event.
Events are stored into a map. A timer is associated with every map
entry. As soon as the timer runs out of time, the corresponding entry
is removed.

Bug: 1352034
Test: tools/autotest.py -C out/Default --run_all ./chrome/browser/ash/policy/dlp/

Report src and dst patterns

When reporting file events, report source and destination patterns
instead of full source and destination.

Bug: 1362452
Test: tools/autotest.py -C out/Default --run_all ./chrome/browser/ash/policy/dlp/
Change-Id: Ic612da49616d8735a657064d09384fd4c7ef05a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3899097
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Luca Accorsi <accorsi@google.com>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055699}