Accessibility: Fix potential UAF in AccessibilityManager
This patch fixes a potential UAF as described in the bug. The issue is
that once CreatePumpkinData is called, the remaining code path can't
be canceled. Even if the user signs out (therefore destroying
AccessibilityManager), CreatePumpkinData and OnPumpkinDataCreated will
still execute. This patch addresses the UAF in two ways:

1. We change CreatePumpkinData from a member to an anonymous function.
Doing so removes any references to members of AccessibilityManager.
2. We bind OnPumpkinDataCreated using a weak pointer, so that the
task is dropped if AccessibilityManager is destroyed.

AX-Relnotes: N/A
Fixed: 1366415
Change-Id: I5c5f31b1a9a0d3cd33b670b1095a8a4fe36f13b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3936206
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055762}