[lacros] Drop the lacros exit wait time from 3s to 2.5s
This CL drops the time that ash waits for lacros-chrome to
successfully exit after BrowserManager receives the shutdown signal
from 3s to 2.5s.

The session_manager will wait for ash to exit successfully for 3s
before forcing the shutdown with a SIGABRT.

Lacros-chrome is sent a SIGTERM by ash as soon as ash shutdown
is detected. However the session_manager's timeout is started
before the wait for lacros-chrome.

This means that, with the current timeout, if lacros-chrome does
not successfully exit before the 3s wait, ash will be killed by
the session_manager and the session will be incorrectly recorded
as an ash-crash. In such cases ash should instead be recorded as
a clean exit and lacros-chrome should be recorded as a crash.

By dropping the time ash waits for lacros-chrome to exit to 2.5s,
the timeout is kept below session_manager's timeout and ash will
be allowed to exit successfully if lacros takes too long to exit.

Bug: 1351479
Change-Id: I1d230670406306c53b8ded068f4bfc8df214b7f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3934428
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055779}