[GPU] Mark StreamTextureSII's TexturePassthrough as bound when binding
https://chromium-review.googlesource.com/c/chromium/src/+/3790843
changed VideoSurfaceTextureImageBacking to correctly mark the GLImage as
bound when binding, thus avoiding the GL decoders' thinking that they
need to lazily bind the image. However, it did so only in the flow where
the validating decoder is being used.

This CL adds such marking in the case where the passthrough decoder is
being used.

Note: This change is not relevant in the case where we pass null for the
GLImage when invoking BindStreamTextureImage(), as in that case lazy
binding will not be performed due to the GLImage not being present.
While that case is currently the default on trunk, there is still a
chance that we will have to disable it via a killswitch. We are adding
this code to ensure that if we do have to disable it, we don't see a
regression in lazy binding occurring on Android. When we definitively
eliminate needing to pass GLImage in BindStreamTextureImage(), we will
also eliminate the code in VideoSurfaceTextureImageBacking that is
marking the image as bound (for both the validating and passthrough
decoders).

Bug: 1323341
Change-Id: I0a10ecbb289bc2bc64227c01025e0c9ad89ed7b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3930919
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055812}