WebUI: QuickSettings: Portal network shows signin on click, update text
This CL updates the quick settings to show the portal sign in when a
network list item is clicked that is in a portal state. Along with this,
the message is updated depending on if the network is in a portal state
or not.

This CL also updated the NetworkRowClickedAction to include a metric for
opening the portal signin from the list of networks in the quick
settings.

Screenshots:
1. No Internet or Portal Suspected subtext:
https://screenshot.googleplex.com/xCD29YTZLoTpKEn.png
2. Portal or Proxy Auth subtext:
https://screenshot.googleplex.com/6Y3oPsHKmKvNqYQ.png
3. Clicking on Portal subtext network list item:
https://screenshot.googleplex.com/6PmRxrpKHF7qDwi

BUG=b:244353766
TEST=ash_unittests --gtest_filter="*NetworkDetailedViewController*",
ash_unittests --gtest_filter="*NetworkListNetworkItemView*"

Change-Id: Ifa9823a0e71c4929f165233dd6bcf13051ee9853
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3930732
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Andre Le <leandre@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Michael Rygiel <michaelrygiel@google.com>
Cr-Commit-Position: refs/heads/main@{#1055816}