Fix flaky test CdmStorageTest.ParallelRead and update ParallelWrite
The test CdmStorageTest.ParallelRead is flaking as both reads into the
CDM file are succeeding intermittently. The purpose of this test is to
test the locking mechanism, but it makes more sense just to see if the
first read succeeds, and the second read we can let it fail due to our
locking mechanism, or succeed if the first read finishes faster than
expected. To keep consistency, this CL also changes the test
CdmStorageTest.ParallelWrite to also just check the first write.

Bug: 1302616
Change-Id: I02520c10bd150be65a69cf260599963d839caa27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3928414
Reviewed-by: Austin Sullivan <asully@chromium.org>
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055835}