Avoid restoring opacity for AppListPage
When the app list changes between different AppListViewStates, the
opacity is updated according to the new state.

When dragging the app list, the opacity of the pages and elements could
be restored to improve the visual of the animation. However, since the
app list drag became obsolete with the introduction of the productivity
launcher, restoring opacity is no longer relevant as a parameter.

This CL removes the parameter in AppListPage and the derived methods
that depend on it.

Bug: 1356674
Change-Id: I6bb843d1f069a54e83c875d76bd6babf0f580be6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3933274
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055836}