Further restrict implicit value conversions to base::expected.
Only allow implicit conversions from U to base::expected<T, E> if U is
*not* implicitly convertible to E. This makes it harder to write bugs of
the form:

  // `e` contains an expected value of `true`, not an unexpected value
  // of `123L`.
  base::expected<bool, long> e = 123L;

As a Chromium extension to the C++23 std::expected proposal, this also
adds a `base::ok` hint to simplify construction of a `base::expected`
containing a value when the implicit conversion is disallowed.

While the C++23 proposal chose not to go this route, section 3.7 calls
a hypothetical version of this `success`. Chromium chooses to call this
`base::ok` because:
- it is succinct (every column is precious when there's only 80!)
- has precedence, e.g. Rust's Result enum uses Ok.

Bug: 1369769
Change-Id: Id6ab0e1f30441c6084a85869dc989fe275aca7de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3929368
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055871}