Favicon: Add a flag to force light mode
This CL adds an optional flag `forceLightMode` to chrome://favicon2 URLs
so that UIs that do not have a dark mode version can get light mode
default favicons. Specifically, the realbox on the NTP is displayed in
light mode even if the NTP and the browser are in dark mode.

Bug: 1279611
Change-Id: I01b15d8bcd41961afa637bd85ef883c4c3858492
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3935920
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055899}