Change policy_modifications table column back to "site_owner"
The column was renamed to "primary_site" in crrev.com/c/3928738. But
it is causing some issue (crbug.com/1371545) that blocks the merge of
crrev.com/c/3857668 which makes method calls to read/write from/to
this table. Thus the plan is to revert the renaming change first,
merge the crrev.com/c/3857668, and then try to rename the column in
a followup CL.

Bug: 1219656, 1371545
Change-Id: I1b55932b8ee18e4495e497e76b195143fe3565fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3936337
Reviewed-by: Chris Fredrickson <cfredric@chromium.org>
Commit-Queue: Shuran Huang <shuuran@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055913}